+++ src/config/logger.config.ts
@@ -60,8 +60,14 @@ export class Logger {
   private readonly configService = configService;
   constructor(private context = 'Logger') {}
 
+  private instance = null;
+
   public setContext(value: string) {
     this.context = value;
+  }
+
+  public setInstance(value: string) {
+    this.instance = value;
   }
 
   private console(value: any, type: Type) {
@@ -75,6 +81,8 @@ export class Logger {
         console.log(
           /*Command.UNDERSCORE +*/ Command.BRIGHT + Level[type],
           '[Evolution API]',
+          Command.BRIGHT + Color[type],
+          this.instance ? `[${this.instance}]` : '',
           Command.BRIGHT + Color[type],
           `v${packageJson.version}`,
           Command.BRIGHT + Color[type],
@@ -99,6 +107,7 @@ export class Logger {
       } else {
         console.log(
           '[Evolution API]',
+          this.instance ? `[${this.instance}]` : '',
           process.pid.toString(),
           '-',
           `${formatDateLog(Date.now())}  `,