+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 1.7.3 (2024-04-18 12:07)
+
+### Fixed
+* Revert fix audio encoding
+* Recovering messages lost with redis cache
+* Adjusts in redis for save instances
+* Adjusts in proxy
+* Revert pull request #523
+* Added instance name on logs
+* Added support for Spanish
+* Fix error: invalid operator. The allowed operators for identifier are equal_to,not_equal_to in chatwoot
+
 # 1.7.2 (2024-04-12 17:31)
 
 ### Feature